lumen: move server logic to server.rs
